requirements.txt
numpy
numpy-stl
pandas
scipy
matplotlib
jupyter
matplotlib
torch
pyvista
trimesh
tqdm
imageio